Import documentation from the docs repo to README
Import relevant bits from
https://github.com/ixofoundation/docs/blob/master/developers/api.md
into the README.

Edit the imported bits heavily to make the resulting documentation
more concise than the original.

Documenting this library in an external repo is not a very good
idea anyway. Documentations should always be bundled with the
libraries unless there is a convincing reason.